Complete portal function
Tips：I have replaced the rendering pipeline with URP, so some of your materials have also been upgraded to URP compliant materials by me. Therefore, it is possible that the game object material has turned pink. The solution is to set the material to Standard in the Inspector window!
Achievement：
Implemented GameManager and SceneController logic; Improved the logic of enemies pursuing players, now enemies will not keep chasing players; The logic implementation of the pop-up window in the map has been completed, and now players can interact with the props on the map and view the content on the "note"